apps/flappy-rl-demo/requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
pygame>=2.5.0
imageio>=2.31.0
Pillow>=10.0.0
